task exception handling + redid test file
added exception handling to basic_task.
without a definition for unhandled_exception() in the promise type, apparently compilers are free to not use a try-catch around the coroutine body.
this means the final await will not be triggered and the task will be destroyed.
this was leading to undefined memory accesses in the new exception test cases.

this is easily fixed by defining unhandled_exception() and forwarding a std::exception_ptr through the wait() method.
thus, this propagates coroutine exception upon .wait() or co_await (which uses .wait()).
additionally, (so long as it's defined) immediately after calling unhandled_exception() the coroutine enters the final suspend and is therefore done().
so most of the code doesn't need to change.

removed the .get() method from basic_task<T> (where T is not void).
this is replaced with .wait() to be more similar to basic_task<void> (which only has .wait()).
also, i didn't like .get() because its name doesn't necessarily imply that it blocks for completion.
i think it's clearer this way.

changed non-void basic_task<T> to use std::variant for holding the T object (and the exception pointer).
this is less error prone because i don't have to concern myself with the lifetime of whatever type it contains, and it (potentially) saves space.
by making std::exception_ptr the first type in the std::variant we have the same behavior as the buffer, where T doesn't have to be default-constructible.

found a typo where basic_task<T>::resume() could resume past the final suspend (was using co.done() instead of !co.done()).

reintroduced the commented-out static_assert in non-void basic_task<T> that ensures T is not a reference type.
if references need to be returned, use std::reference_wrapper<T>.
i might (but probably won't) specialize cases for returning references - if i do it'll likely just be sfinae magic to be more concise.

changed task<T> and lazy_task<T> alias templates to put T through std::remove_cv_t.
this is done so that we're not storing const/volatile objects inside the promise.
this also has the benefit of making task<void> the same as task<const void> etc.
the cv qualifiers of a return type aren't really important anyway.

redid the test file to be shortened by macros and also not print anything to the screen unless there's an error (failed test case).
test cases now use co_routine lambdas to be even more concise (they're so pretty...).