added basic_generator
added the basic_generator class.
it has an experimental feature that i'm probably going to remove later on (just want it safe in a commit somewhere):
the iterator (nested) class upon increment returns a pseudo_iterator object, which performs the work of doing the actual increment.
the increment is performed either asynchronously via supporting co_await semantics or synchronously when destroyed, whichever comes first.
the benefit of this is that it could theoretically work both for normal ++it and co_await ++it (as a for co_await statement would translate to).
however the support for the async form is not entirely completed - i don't really feel like completin it because i know i'm going to scrap it anyway.
the issue is that the result of ++it could be stored in an auto variable declaration, which means the destructor isn't called when i expect it, so the increment is subverted.
additionally, it's not kosher to return something other than iterator& from an iterator increment overload.

i'm planning on making basic_generator take additional template parameters to say whether it should be sync or async.
this will be wrapped in generator<T> and async_generator<T> (or something similar) whenever it's added.

added the is_task and is_task_v type traits for determining if some type is a (potentially cv-qualified) basic_task type.
added wait_all() and wait_any() - they only participate in overload resolution if all their (param pack) args at task types.

changed the custom assertion macros to also write the line number of the assertion failure.