task cleanup + reference-returning tasks
merged the void and non-void forms of basic_task into one class definition.
the only differences are the promise type, which is left as a separate specialized template class in the detail namespace, and the wait() implementation.
constexpr bool flags are (privately) defined to differentiate the different forms of basic_task.
these are then used inside constexpr if's to implement wait() in a very easy-to-read manner.

since basic_task is nice and simple to extend now, added lvalue and rvalue reference forms (in addition to the void and value forms).
additionally both forms of reference can share the same wait() code (but not the same promise code due to different return_value() arg types).
added a bunch of test cases for making sure wait() and co_await return the expected types (reference categories are preserved).

changed the semantics of wait():
it now returns a (move-constructed) value (except in the reference and void returning cases).
and to guarantee there are no accidental accesses after this point, the task enters the empty state after this.
so subsequent calls to wait() (e.g. to get the value again) will throw bad_coroutine_access.
wait() and resume() are now both changed to be non-const methods (were only originally const because coroutine_handle<>::resume() is a const method).
with the new wait() semantics, the rvalue overload of wait() is no longer needed - yay.